Hide internal api types and add presences parameter to Send match data.

+++ nakama/example/lib/widgets/match_area.dart
@@ -39,7 +39,7 @@ class MatchAreaState extends State<MatchArea> {
     // Send dummy match data via Websocket
     NakamaWebsocketClient.instance.sendMatchData(
       matchId: widget.match.matchId,
-      opCode: Int64(0),
+      opCode: 0,
       data: data.codeUnits,
     );
     print('Match Data changed: $data');
